[+] allow `RowSet` with error with no corresponding row (#245)
Allow a RowSet to be specified with an error row number that has no
corresponding row. Since the row returns an error, it seems unnecessary
for there to be a row.

In other words, rather than require something like:

	NewRows([]string{"id", "title"}).
		AddRow("one").
		RowError(0, fmt.Errorf("error"))

In which both a row and a row error specify index 0, but the row is
never returned, allow omission of the row:

	NewRows([]string{"id", "title"}).
		RowError(0, fmt.Errorf("error"))

The pattern was broken in d07ccec, which checked only the number of rows
before closing a RowSet. Fix by having `Next` check both the length of
rows and whether an error is registered for the current row number. Add
a test for this case, as well.